+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.10.2] - 2026-04-08
+
 ### Fixed
 
 - `list_messages` now annotates the `FROM` column with `(via <domain>)`